% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gambit.r
\name{cond.eq.outcomes}
\alias{cond.eq.outcomes}
\title{return a conditional equilibrium outcome
cond is a list with variable names and their assumed value
we only pick rows from oco.df in which the condition is satisfied
we set the probabilities of the conditioned variable values to 1}
\usage{
cond.eq.outcomes(eq.li, cond, tg = NULL, oco.df = tg$oco.df)
}
\description{
return a conditional equilibrium outcome
cond is a list with variable names and their assumed value
we only pick rows from oco.df in which the condition is satisfied
we set the probabilities of the conditioned variable values to 1
}
