% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game.r
\name{get.project.dir}
\alias{get.project.dir}
\title{Get the gtree project directory}
\usage{
get.project.dir(xs = get.xs())
}
\description{
By default the current working directory
but the gtree GUI can set a different folder
}
